青い鳥が追加されてるよ

+++ Project1/Manage/Menu.cpp
@@ -239,11 +239,14 @@ bool Menu::Mix(ItemSave& item1, ItemSave& item2, ItemSave& item3)
 	if ((item1.itemtype == ITEM_TYPE::BOOK || item2.itemtype == ITEM_TYPE::BOOK)&& 
 		(item1.itemtype == ITEM_TYPE::STONE || item2.itemtype == ITEM_TYPE::STONE))
 	{
-		if (Key(item1, item2)){	return true;}
-		else if (Vine(item1, item2)){return false;}
+		if (Key(item1, item2)) { return true; }					// 鍵のレシピ
+		else if (Vine(item1, item2)) { return true; }			// 蔓のレシピ
+		else if (Dynamite(item1, item2)) { return true; }		// ダイナマイトのレシピ
+		else if (Hose(item1, item2)) { return true; }			// ホースのレシピ
+		else if (Bard(item1, item2)) { return true; }			// 鳥のレシピ
 		else{return false;}
 	}
-	else if (item1.itemtype == ITEM_TYPE::BOOK || item2.itemtype == ITEM_TYPE::BOOK)
+	else if (item1.itemtype != ITEM_TYPE::STONE || item2.itemtype != ITEM_TYPE::STONE)
 	{
 		return false;
 	}
@@ -296,6 +299,72 @@ bool Menu::Vine(ItemSave& item1, ItemSave& item2)
 			(item2.itemtype == ITEM_TYPE::BOOK && item2.colortype == COLOR_TYPE::GREEN))
 		{
 			item1.colortype = COLOR_TYPE::GREEN;
+			item1.itemtype = ITEM_TYPE::TOOL;
+			lpTradeMng.AddBag(item1);
+			lpTradeMng.DeleteItem(item2);
+			for (int i = 0;i < _asize;i++)
+			{
+				_sItem.at(i).first.itemtype = ITEM_TYPE::NON;
+			}
+			return true;
+		}
+	}
+	return false;
+}
+
+bool Menu::Dynamite(ItemSave& item1, ItemSave& item2)
+{
+	if ((item1.itemtype == ITEM_TYPE::STONE && item1.colortype == COLOR_TYPE::RED) ||
+		item2.itemtype == ITEM_TYPE::STONE && item2.colortype == COLOR_TYPE::RED)
+	{
+		if ((item1.itemtype == ITEM_TYPE::BOOK && item1.colortype == COLOR_TYPE::RED) ||
+			(item2.itemtype == ITEM_TYPE::BOOK && item2.colortype == COLOR_TYPE::RED))
+		{
+			item1.colortype = COLOR_TYPE::RED;
+			item1.itemtype = ITEM_TYPE::TOOL;
+			lpTradeMng.AddBag(item1);
+			lpTradeMng.DeleteItem(item2);
+			for (int i = 0;i < _asize;i++)
+			{
+				_sItem.at(i).first.itemtype = ITEM_TYPE::NON;
+			}
+			return true;
+		}
+	}
+	return false;
+}
+
+bool Menu::Hose(ItemSave& item1, ItemSave& item2)
+{
+	if ((item1.itemtype == ITEM_TYPE::STONE && item1.colortype == COLOR_TYPE::CYAN) ||
+		item2.itemtype == ITEM_TYPE::STONE && item2.colortype == COLOR_TYPE::CYAN)
+	{
+		if ((item1.itemtype == ITEM_TYPE::BOOK && item1.colortype == COLOR_TYPE::BLUE) ||
+			(item2.itemtype == ITEM_TYPE::BOOK && item2.colortype == COLOR_TYPE::BLUE))
+		{
+			item1.colortype = COLOR_TYPE::CYAN;
+			item1.itemtype = ITEM_TYPE::TOOL;
+			lpTradeMng.AddBag(item1);
+			lpTradeMng.DeleteItem(item2);
+			for (int i = 0;i < _asize;i++)
+			{
+				_sItem.at(i).first.itemtype = ITEM_TYPE::NON;
+			}
+			return true;
+		}
+	}
+	return false;
+}
+
+bool Menu::Bard(ItemSave& item1, ItemSave& item2)
+{
+	if ((item1.itemtype == ITEM_TYPE::STONE && item1.colortype == COLOR_TYPE::BLUE) ||
+		item2.itemtype == ITEM_TYPE::STONE && item2.colortype == COLOR_TYPE::BLUE)
+	{
+		if ((item1.itemtype == ITEM_TYPE::BOOK && item1.colortype == COLOR_TYPE::BLUE) ||
+			(item2.itemtype == ITEM_TYPE::BOOK && item2.colortype == COLOR_TYPE::BLUE))
+		{
+			item1.colortype = COLOR_TYPE::BLUE;
 			item1.itemtype = ITEM_TYPE::TOOL;
 			lpTradeMng.AddBag(item1);
 			lpTradeMng.DeleteItem(item2);
